requirements.txt
Cython==3.0.10
matplotlib==3.9.0
numpy==1.26.4
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
point3d==0.1
pyntcloud==0.3.1
setuptools==70.0.0
torch==2.3.0
vispy==0.14.2
